LPS-77585 Prepare extracting web only methods to internal web ShoppingUtil class.
